+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - [Update]: This version adds a new `OTSession.forceDisconnect()` method for forcing clients to disconnect.
 
-- [Update]: This version adds new OTPublisher `allowAudioCaptureWhileMuted` property of the OTPublisher `props` object. Set this to `true` to control audio capture while muted so the publisher receives audio level updates when not publishing audio. This is useful for implemeting features such as speaking while muted.
+- [Update]: This version adds an `allowAudioCaptureWhileMuted` property of the OTPublisher `properties` prop. Set this to `true` to have the publisher receive audio level updates when not publishing audio.
 
 # 2.31.0-beta.0 (July 2025)